# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/subscription:
    get:
      summary: 'Get active subscription'
      operationId: getActiveSubscription
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  start: 1686580275
                  start_date: '2023-01-02 10:00:12'
                  end: 1686580275
                  end_date: '2023-01-02 11:00:12'
                properties:
                  start:
                    type: integer
                    example: 1686580275
                  start_date:
                    type: string
                    example: '2023-01-02 10:00:12'
                  end:
                    type: integer
                    example: 1686580275
                  end_date:
                    type: string
                    example: '2023-01-02 11:00:12'
      tags:
        - Endpoints
      security: []
  /api/current-stats:
    get:
      summary: 'Get current stats'
      operationId: getCurrentStats
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  tasks: 18
                  can_add_task: true
                  projects: 2
                  can_add_project: false
                  space: 512012
                  can_add_files: false
                properties:
                  tasks:
                    type: integer
                    example: 18
                  can_add_task:
                    type: boolean
                    example: true
                  projects:
                    type: integer
                    example: 2
                  can_add_project:
                    type: boolean
                    example: false
                  space:
                    type: integer
                    example: 512012
                  can_add_files:
                    type: boolean
                    example: false
      tags:
        - Endpoints
      security: []
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []